git: optimize checkout
only do the checkout and repo clean if the current HEAD is different
from the target commit.

+++ jolt/plugins/git.py
@@ -33,6 +33,7 @@ class GitRepository(object):
         self.refspecs = refspecs or []
         self._tree_hash = {}
         self._original_head = True
+        self._last_rev = None
         self._init_repo()
 
     def _init_repo(self):
@@ -215,17 +216,22 @@ class GitRepository(object):
                 output_on_error=True)
 
     def checkout(self, rev, commit=None):
+        if rev == self._last_rev:
+            log.debug("Checkout skipped, already @ {}", rev)
+            return False
         log.info("Checking out {0} in {1}", rev, self.path)
         with self.tools.cwd(self.path):
             try:
-                return self.tools.run("git checkout -f {rev}", rev=rev, output=False)
+                self.tools.run("git checkout -f {rev}", rev=rev, output=False)
             except Exception:
                 self.fetch(commit=commit)
                 try:
-                    return self.tools.run("git checkout -f {rev}", rev=rev, output_on_error=True)
+                    self.tools.run("git checkout -f {rev}", rev=rev, output_on_error=True)
                 except Exception:
                     raise_error("Commit does not exist in remote for '{}': {}", self.relpath, rev)
         self._original_head = False
+        self._last_rev = rev
+        return True
 
 
 _gits = {}
@@ -370,9 +376,9 @@ class GitSrc(WorkspaceResource, FileInfluence):
             # Should be safe to do this now
             rev = self.git.rev_parse(rev)
             if not self.git.is_head(rev) or self._revision.is_imported:
-                self.git.checkout(rev, commit=commit)
-                self.git.clean()
-                self.git.patch(self._diff.value)
+                if self.git.checkout(rev, commit=commit):
+                    self.git.clean()
+                    self.git.patch(self._diff.value)
 
     def get_influence(self, task):
         return None
